move wallet mocks to chain implementation files
