+++ packages/chain-ethereum/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./signatureData.js"
 export * from "./implementation.js"
+export * from "./walletMock.js"
