+++ packages/chain-substrate/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./implementation.js"
+export * from "./walletMock.js"
